Update version of logger script
Contains fixes (including a namespace declaration, which was the
previous problem. It's also been moved to it's own GitHub repo, so that
I can get proper revision hashes for each version. (Gist seems to keep
the same hash for the latest version of the file? This would break
upgrades to the script in the future).

This "more works" but still getting unhelpful feedback from WP-CLI, so
more to do.