Wokarounds an issue with the Selenium driver where some elements can be hidden by the toolbar
The 'admin' toolbar that appears when logged in, can hover over elements
that we want to interact with. If such an element is underneath the toolbar
then Selenium does not allow us interact with it.

'Normal' user behaviour would be to move the view port up or down, but this
is difficult to achieve. Instea, we use JavaScript to remove the 'fixed'
nature of the toolbar, so that it sits at the top of the page. When in the
wp-admin/ some additional positioning is required to ensure the toolbar
does not cover any content.

For Goutte driver this workaround is not required.

Fixes #81