If a browser window is not open when the BeforeStep event is run then our call to Selenium2Driver::executeScript() will throw an 'unknown' exception.
This will typically happen on the first step if there are no Background steps which open a window (BeforeStep does not
run before Background steps).

In such situations it is of no consequence that the script is not run, since it is a workaround for the admin toolbar
hiding clickable elements which is only an issue when we are interacting with the page via the browser.